feat(call-room): screen-glow rises from card into dark room (immersive night)
Removed the fake warm light blob painted on the card itself — that read
as a Photoshop dodge artifact, not a 3D light source.

New approach: at night the user's actual laptop/monitor would be the
only thing in the dim virtual office emitting light. Modeled exactly
that — a cool-blue screen-glow that rises FROM the card's top edge UP
into the dark room above. Two stacked radial ellipses with screen
blend, anchored at 50% horizontal, 28-32% vertical (above where the
card top sits):

  1. Tall narrow ellipse (36%×28% at 50%/28%) — concentrated bright
     zone where the 'screen' meets the room. rgba(170,195,235,0.30)
     core fading to transparent at 80%.
  2. Wider softer halo (65%×45% at 50%/32%) — diffuse spread so the
     whole upper-room area picks up gentle screen-light, not just a
     spotlight pinned above the card.

The card itself stays clean glass — no painted light on its surface.
Light now appears to emanate FROM the card UP into the room above.
Crossfades 0→1 over 800ms in sync with the night image swap.

Refactor:
  - CallRoomCard no longer takes forcedTimeOfDay (was only used for
    the now-removed card-light blob). Kept the prop for compat.
  - CallRoom.tsx now calls useTimeOfDay and renders the screen-glow
    between background and cardWrap.

Tests: 29/29 still pass.
